+++ common/src/main/java/com/yanny/ali/plugin/client/widget/TagWidget.java
@@ -3,11 +3,10 @@ package com.yanny.ali.plugin.client.widget;
 import com.yanny.ali.api.*;
 import com.yanny.ali.plugin.common.nodes.TagNode;
 
-public class TagWidget extends IWidget {
+public class TagWidget implements IWidget {
     private final RelativeRect bounds;
 
     public TagWidget(IWidgetUtils utils, IDataNode entry, RelativeRect rect, int maxWidth) {
-        super(entry.getId());
         TagNode tagEntry = (TagNode) entry;
 
         utils.addSlotWidget(tagEntry.getModifiedItem(), tagEntry, rect);