+++ common/src/main/java/com/yanny/ali/plugin/client/widget/TextureWidget.java
@@ -14,7 +14,7 @@ import net.minecraft.resources.ResourceLocation;
 import java.util.LinkedList;
 import java.util.List;
 
-public class TextureWidget extends IWidget {
+public class TextureWidget implements IWidget {
     protected final ResourceLocation texture;
     protected final RelativeRect rect;
     protected final int u;
@@ -26,7 +26,6 @@ public class TextureWidget extends IWidget {
     private final List<Component> components = new LinkedList<>();
 
     public TextureWidget(ResourceLocation texture, RelativeRect rect, int u, int v, int regionWidth, int regionHeight, int textureWidth, int textureHeight) {
-        super(ResourceLocation.withDefaultNamespace("texture_widget"));
         this.texture = texture;
         this.rect = rect;
         this.u = u;