+++ fabric/src/main/java/com/yanny/ali/network/NetworkUtils.java
@@ -21,9 +21,11 @@ public class NetworkUtils {
         Server server = new Server();
 
         PayloadTypeRegistry.playS2C().register(SyncLootTableMessage.TYPE, SyncLootTableMessage.CODEC);
+        PayloadTypeRegistry.playS2C().register(SyncTradeMessage.TYPE, SyncTradeMessage.CODEC);
         PayloadTypeRegistry.playS2C().register(ClearMessage.TYPE, ClearMessage.CODEC);
         PayloadTypeRegistry.playS2C().register(DoneMessage.TYPE, DoneMessage.CODEC);
         ClientPlayNetworking.registerGlobalReceiver(SyncLootTableMessage.TYPE, client::onLootInfo);
+        ClientPlayNetworking.registerGlobalReceiver(SyncTradeMessage.TYPE, client::onTradeInfo);
         ClientPlayNetworking.registerGlobalReceiver(ClearMessage.TYPE, client::onClear);
         ClientPlayNetworking.registerGlobalReceiver(DoneMessage.TYPE, client::onDone);
         return new DistHolder<>(client, server);
@@ -34,6 +36,7 @@ public class NetworkUtils {
         Server server = new Server();
 
         PayloadTypeRegistry.playS2C().register(SyncLootTableMessage.TYPE, SyncLootTableMessage.CODEC);
+        PayloadTypeRegistry.playS2C().register(SyncTradeMessage.TYPE, SyncTradeMessage.CODEC);
         PayloadTypeRegistry.playS2C().register(ClearMessage.TYPE, ClearMessage.CODEC);
         PayloadTypeRegistry.playS2C().register(DoneMessage.TYPE, DoneMessage.CODEC);
         return new DistHolder<>(null, server);