+++ forge/src/main/java/com/yanny/ali/network/Server.java
@@ -17,7 +17,12 @@ public class Server extends AbstractServer {
     }
 
     @Override
-    protected void sendSyncMessage(ServerPlayer serverPlayer, SyncLootTableMessage message) {
+    protected void sendSyncLootTableMessage(ServerPlayer serverPlayer, SyncLootTableMessage message) {
+        channel.send(message, PacketDistributor.PLAYER.with(serverPlayer));
+    }
+
+    @Override
+    protected void sendSyncTradeMessage(ServerPlayer serverPlayer, SyncTradeMessage message) {
         channel.send(message, PacketDistributor.PLAYER.with(serverPlayer));
     }