+++ neoforge/src/main/java/com/yanny/ali/network/NetworkUtils.java
@@ -21,6 +21,7 @@ public class NetworkUtils {
         Server server = new Server();
 
         registrar.executesOn(HandlerThread.NETWORK).playToClient(SyncLootTableMessage.TYPE, SyncLootTableMessage.CODEC, client::onLootInfo);
+        registrar.executesOn(HandlerThread.NETWORK).playToClient(SyncTradeMessage.TYPE, SyncTradeMessage.CODEC, client::onTradeInfo);
         registrar.executesOn(HandlerThread.NETWORK).playToClient(ClearMessage.TYPE, ClearMessage.CODEC, client::onClear);
         registrar.executesOn(HandlerThread.NETWORK).playToClient(DoneMessage.TYPE, DoneMessage.CODEC, client::onDone);
         return new DistHolder<>(client, server);
@@ -31,6 +32,7 @@ public class NetworkUtils {
         Server server = new Server();
 
         registrar.playToClient(SyncLootTableMessage.TYPE, SyncLootTableMessage.CODEC, (a, b) -> {});
+        registrar.playToClient(SyncTradeMessage.TYPE, SyncTradeMessage.CODEC, (a, b) -> {});
         registrar.playToClient(ClearMessage.TYPE, ClearMessage.CODEC, (a, b) -> {});
         registrar.playToClient(DoneMessage.TYPE, DoneMessage.CODEC, (a, b) -> {});
         return new DistHolder<>(null, server);