+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## []
+
+- Added Trade info support for vanilla
+- Fixed memory leak in JEI plugin
+
+## [1.2.5]
+
+- Fixed no LootJS info displayed when adding to empty loot table
+
 ## [1.2.4]
 
 - Fixed no tooltip in part of right panel in REI